requirements.txt
kafka-python==1.4.5 
elasticsearch>=1.0.0,<2.0.0
python-twitter